+++ SecureHttpClient.Test/HttpTest.cs
@@ -149,7 +149,7 @@ namespace SecureHttpClient.Test
         [SkippableFact]
         public async Task HttpTest_DeleteCookie()
         {
-            Skip.If(DeviceInfo.Platform == DevicePlatform.Android && DeviceInfo.Version.Major == 7, "Failing on Android 24-25");
+            Skip.If((DeviceInfo.Platform == DevicePlatform.Android && DeviceInfo.Version.Major == 7) || DeviceInfo.Platform == DevicePlatform.UWP, "Failing on Android 24-25 and on UWP");
 
             const string page1 = @"https://httpbin.org/cookies/set?k1=v1";
             await GetAsync(page1);